Actaulizar usuario

+++ src/main/java/ec/edu/espe/gpr/dao/IDocenteDao.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ec.edu.espe.gpr.model.Docente;
 import ec.edu.espe.gpr.model.Usuario;
 
-public interface IDocenteDao extends JpaRepository<Docente, Long> {
+public interface IDocenteDao extends JpaRepository<Docente, Integer> {
     Optional<Docente> findByCodigoDocente(Integer codigoDocente);
     Optional<Docente> findByCodigoUsuario(Usuario codigoUsuario);
 }
